+++ src/extension.ts
@@ -2,6 +2,7 @@ import * as fs from "fs"
 import * as path from 'path'
 import * as vscode from 'vscode'
 import Message from './common/Message'
+import i18nConstructor from "./util/i18n";
 import Config from './Config'
 import Terminal from './common/Terminal'
 import Command from './Command'
@@ -9,7 +10,8 @@ import Command from './Command'
 const dataPath = path.join(__dirname, '../../data.json')
 
 let message = new Message('Run As')
-let config = new Config(process.platform, message)
+let i18n = new i18nConstructor('en', vscode.env.language, path.join(__dirname, '../../locale/lang.%s.json'))
+let config = new Config(process.platform, message, i18n)
 let terminal = new Terminal('Run As ...')
 
 if (isFirstRun()) remindUpdateConfig()
@@ -17,10 +19,10 @@ if (isFirstRun()) remindUpdateConfig()
 export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(config.reloadOnConfigChange())
     context.subscriptions.push(terminal.initOnClose())
-    context.subscriptions.push(vscode.commands.registerCommand('extension.runas', e => {
+    context.subscriptions.push(vscode.commands.registerCommand('extension.runAs', e => {
         if (e.fsPath) {
             let commandExpression = config.getCommandMapByFile(e.fsPath),
-                command = new Command(commandExpression, e.fsPath, config.newWindowConfig as any, message)
+                command = new Command(commandExpression, e.fsPath, config.newWindowConfig as any, message, i18n)
             terminal.exec(command.toString())
         }
     }))
@@ -33,6 +35,6 @@ function isFirstRun() {
 }
 
 function remindUpdateConfig() {
-    message.info('If you has config this extension configuration, you need to modify the user configuration to make sure it work correctly.')
+    message.info(i18n.get('info.firstRun'))
     fs.writeFileSync(dataPath, JSON.stringify({ isFirstRun: false }))
 }